DCADemo/app/requirements.txt
Flask
mysql-connector
pandas
pymc3